+++ assets/index.js
@@ -3,6 +3,7 @@ export * from "./images/list/index";
 export * from "./images/login/index";
 export * from "./images/chat/index";
 export * from "./colors";
+export * from "./images/my-page/index";
 
 const defaultProfileImage = require("./images/default_profile_image.png");
 export { defaultProfileImage };
